Fix buffer size: 64bit GLBufferSizeTracker, hash over addr+size, save fail-fast hash cache, ..
Relates to GlueGen 6b6b9b3b81cdc85b7260664ebec547756a6be5d7, branch sgothel_wip_fixes01.

Memory object size is ptrdiff_t, hence long (64bit).
The hash value must include size as well, otherwise boundaries cannot be verified. (security)
Double check hash collisions while adding a new MemoryObject.

+++ make/config/jogl/gl-impl-CustomCCode-gl4bc.c
@@ -30,10 +30,10 @@ Java_com_jogamp_opengl_impl_gl4_GL4bcImpl_dispatch_1glMapNamedBufferEXT(JNIEnv *
 
 /*   Java->C glue code:
  *   Java package: com.jogamp.opengl.impl.gl4.GL4bcImpl
- *    Java method: ByteBuffer newDirectByteBuffer(long addr, int capacity);
- *     C function: jobject newDirectByteBuffer(jlong addr, jint capacity);
+ *    Java method: ByteBuffer newDirectByteBuffer(long addr, long capacity);
+ *     C function: jobject newDirectByteBuffer(jlong addr, jlong capacity);
  */
 JNIEXPORT jobject JNICALL
-Java_com_jogamp_opengl_impl_gl4_GL4bcImpl_newDirectByteBuffer(JNIEnv *env, jobject _unused, jlong addr, jint capacity) {
+Java_com_jogamp_opengl_impl_gl4_GL4bcImpl_newDirectByteBuffer(JNIEnv *env, jobject _unused, jlong addr, jlong capacity) {
   return (*env)->NewDirectByteBuffer(env, (void*) (intptr_t) addr, capacity);
 }
